requirements.txt
# Core Data Processing
pyspark==3.4.1
pandas==2.0.3
numpy==1.24.3
pyarrow==12.0.1

# Azure SDKs
azure-storage-blob==12.17.0
azure-storage-file-datalake==12.12.0
azure-identity==1.13.0
azure-keyvault-secrets==4.7.0
azure-cosmos==4.4.0
azure-eventhub==5.11.3

# Azure AI Services
azure-search-documents==11.4.0
openai==1.3.0
azure-ai-textanalytics==5.3.0

# Data Validation & Quality
great-expectations==0.17.15
pydantic==2.4.0

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3
requests==2.31.0
jsonschema==4.19.0

# Logging & Monitoring
opencensus-ext-azure==1.1.9
applicationinsights==0.11.10